+++ CHANGELOG.md
@@ -50,6 +50,8 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - :tada: **Enhancement** added convertBarSeries(BarSeries, conversionFunction) to BarSeriesUtils.
 - :tada: **Enhancement** added UnstableIndicator.
 - :tada: **Enhancement** added Chainrule.
+- :tada: **Enhancement** added chooseBest() method with parameter orderType in AnalysisCriterion.
+
 
 
 ## 0.13 (released November 5, 2019)